Move CLI code to its own module
